src/main/java/pl/github/DominasPL/JUnit5/Account.java
package pl.github.DominasPL.JUnit5;

public class Account {

    private boolean active;
    private Address defaultDeliveryAddress;

    public void setActive(boolean active) {
        this.active = active;
    }

    public Address getDefaultDeliveryAddress() {
        return defaultDeliveryAddress;
    }

    public void setDefaultDeliveryAddress(Address defaultDeliveryAddress) {
        this.defaultDeliveryAddress = defaultDeliveryAddress;
    }

    public Account() {
        this.active = false;
    }

    public void activate() {
        this.active = true;
    }

    public boolean isActive() {
        return active;
    }


}
